pyproject.toml
[project]
name = "topino"
version = "0.1.0"
description = "Video motion analysis tool"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Michele Cafagna" }
]
dependencies = [
    "ffmpeg-python>=0.2.0",
    "opencv-python>=4.12.0.88",
    "pyarrow>=21.0.0",
    "pydantic>=2.11.9",
    "scipy>=1.16.1",
    "plotly>=6.3.0",
    "pillow>=11.3.0",
    "pandas>=2.3.2",
    "numpy>=2.2.6",
    "streamlit>=1.49.1",
    "streamlit-cropper>=0.3.1",
]

[project.optional-dependencies]
cli = [
    "typer>=0.17.4",
    "rich>=14.1.0",
]
dev = [
    "topino[cli]",
    "mypy>=1.18.1",
    "ruff>=0.13.0",
    "jupyter>=1.1.1",
]

[project.scripts]
topino = "topino.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

# Ignore missing stubs for specific modules
[[tool.mypy.overrides]]
module = [
    "marimo.*",
    "plotly.*",
    "ffmpeg.*",
    "pandas.*",
    "scipy.*"
]
ignore_missing_imports = true
